Move MinimalErrorScreen to compose-ui module
Refactor to improve code organization by moving MinimalErrorScreen from
jvm-app to compose-ui module where it can be shared across platforms.

Changes:
- Create MinimalErrorScreen.kt in compose-ui/src/commonMain
- Remove MinimalErrorScreen function from Main.kt
- Remove error color constants from Main.kt
- Clean up unused imports in Main.kt
- Add proper import for MinimalErrorScreen in Main.kt
